Move test to new home
